Add SD.h library support
Can now change whether the library compiles for the normal Arduino SD
library or for the SdFat library directly.

Also made a bunch of stuff static so it could be accessed by the static
init() in the PiWatch class. This is mostly driven by the fact that I
can't seem to pass function callbacks when things aren't static. There's
got to be a better way... Need to fork this and play with it.